--- a/boxmot/tracker_zoo.py
+++ b/boxmot/tracker_zoo.py
@@ -47,7 +47,8 @@ def create_tracker(tracker_type, tracker_config=None, reid_weights=None, device=
         'botsort': 'boxmot.trackers.botsort.botsort.BotSort',
         'deepocsort': 'boxmot.trackers.deepocsort.deepocsort.DeepOcSort',
         'hybridsort': 'boxmot.trackers.hybridsort.hybridsort.HybridSort',
-        'imprassoc': 'boxmot.trackers.imprassoc.imprassoctrack.ImprAssocTrack'
+        'imprassoc': 'boxmot.trackers.imprassoc.imprassoctrack.ImprAssocTrack',
+        'boosttrack': 'boxmot.trackers.boosttrack.boosttrack.BoostTrack',
     }
 
     # Check if the tracker type exists in the mapping
@@ -60,10 +61,10 @@ def create_tracker(tracker_type, tracker_config=None, reid_weights=None, device=
     tracker_class = getattr(__import__(module_path, fromlist=[class_name]), class_name)
     
     # For specific trackers, update tracker arguments with ReID parameters
-    if tracker_type in ['strongsort', 'botsort', 'deepocsort', 'hybridsort', 'imprassoc']:
+    if tracker_type in ['strongsort', 'botsort', 'deepocsort', 'hybridsort', 'imprassoc', 'boosttrack']:
         tracker_args['per_class'] = per_class
         tracker_args.update(reid_args)
-        if tracker_type == 'strongsort':
+        if tracker_type in ['strongsort', 'boosttrack']:
             tracker_args.pop('per_class')  # per class not supported by
     else:
         tracker_args['per_class'] = per_class