Hello! I've decided to add BoostTrack to boxmot, because I am currently researching this tracker results and can't wait for someone else to do it :smiley: 

There are many comments, and I could forget something important:


###  1. Reproduce BoostTrack tracker results

Originally, after running BoostTrack scripts from original repository with `val_half.json`, i got:

```
{"HOTA": 68.486, "MOTA": 75.508, "IDF1": 81.277}
```

To reproduce these results in boxmot:

1)  Download [val_half](https://drive.google.com/file/d/1QZXuLZq6zp4_DrOB0kCwYkclNEksFBEh/view?usp=drive_link) from my drive or prepare it by yourself (It contains already splitted MOT17 validation part). Folder path should be `val_utils/data/MOT17/val_half`
2) Download ByteTrack ablation model from [here](https://drive.google.com/file/d/1iqhM-6V_r1FpOlOzrdP_Ejshgk0DxOob/view), move it to boxmot folder and rename it to `yolox_x_ablation.pth`
2) run command:

```
python3 tracking/val.py --yolo-model yolox_x_ablation.pth --reid-model osnet_x1_0_msmt17.pt --tracking-method boosttrack --verbose --source ./tracking/val_utils/data/MOT17/val_half/ --conf 0.1
```

Results should be:

```
{"HOTA": 68.382, "MOTA": 75.558, "IDF1": 81.154}`
```

For closest results you need to add these lines to [val.py](https://github.com/mikel-brostrom/boxmot/blob/dbce5f2bc4eaecbdce702c600ed0cc04fbe6e43b/tracking/val.py#L277) (between `embs = frame_dets_n_embs[:, 7:]` and `tracker.update`):

```
if args.tracking_method == "boosttrack":
            with open(source.parent / "seqinfo.ini") as f: 
                seq_data = f.readlines()

            x = next(sd for sd in seq_data if sd.startswith("frameRate"))
            frame_rate = int(x[x.find("=") + 1:x.find('\n')])

            tracker.max_age = max(30, frame_rate * 2)
```

Idea: authors of boosttrack added [dynamic](https://github.com/vukasin-stanojevic/BoostTrack/blob/8e80cd263ec33144db91c356d194bfcdaa9894bf/default_settings.py#L64) `max_age` a.k.a. `max_time_lost` to their tracker, which uses `max(30, frame_rate * 2)` of each MOT17, MOT20 sequence.

After this change results should be:

```
{"HOTA": 68.517, "MOTA": 75.55, "IDF1": 81.398}
```

I could also reproduce the original result one to one, but it requires too many steps to show it here. The most important thing is that the tracker works exactly the same as the original, the only difference here is in the processing of detections, which is a bit different in boxmot and boosttrack original repo.

### 2. Comparison: BoostTrack vs ByteTrack

In my experience, this tracker shows speed like 70-75 FPS, while bytetrack works on 200 FPS. Main problem is of course in ECC. I turned it off to check:

BoostTrack without ecc: `{"HOTA": 67.42, "MOTA": 75.01, "IDF1": 79.34}`, and FPS is almost equal to ByteTrack
ByteTrack: `{"HOTA": 67.619, "MOTA": 78.081, "IDF1": 79.188}`.

So my verdict is that default BoostTrack is essentially the same or even worse than ByteTrack.

### 3. BoostTrack+ and BoostTrack++

1) BoostTrack+  == BoostTrack with ReID. Currently I didn't check this tracker with ReID, I will do it in next iteration.

2) Of course, main contribution of this tracker is BoostTrack++ version, but currently I don't know how to better add both + and ++ to boxmot. There are only few changes in configs, and now there is no functionality in boxmot to add tracker without changing its class while changing only its config file. This thing should be discussed and maybe you have your own opinion.

#1826