@Fleyderer 2025-03-05
I will check BoxMOT CI / tests tomorrow