@mikel-brostrom 2025-03-05
Nice @Fleyderer!

Add the model here:

https://github.com/mikel-brostrom/boxmot/blob/dbce5f2bc4eaecbdce702c600ed0cc04fbe6e43b/.github/workflows/benchmark.yml#L28C9-L28C16

for automatic benchmarking 😄