@Fleyderer 2025-03-06
In tests, there is only one problem with boosttrack: 

```FAILED tests/unit/test_trackers.py::test_tracker_output_size[boosttrack] - assert (1, 8) == (2, 8)```

This happens because `filter_outputs` function of boosttrack filters out one of tracks, which has aspect_ratio 1.61, while aspect_ratio_thresh in boosttrack is 1.6

I think it is not a problem, because it is original logic of this tracker, and strongsort does not pass this test too.