@mikel-brostrom 2025-03-06
Let's change that specific bbox:

[144,212,400,480,0.82,0]

Will do