@mikel-brostrom 2025-03-06
Also add the tracking method name here:

https://github.com/mikel-brostrom/boxmot/blob/6486fc79e0b2d066cfe8a02df9054e0af7cc5f21/.github/workflows/ci.yml#L114