@mikel-brostrom 2025-03-06
The set of detections used for each may not be exactly the same. When i run the benchmark job I get the following:

With ECC
``` 
                                      Tracker	Status	HOTA↑	MOTA↑	IDF1↑	FPS
[botsort](https://arxiv.org/abs/2206.14651)	✅	68.774	77.176	80.997	46
[strongsort](https://arxiv.org/abs/2202.13514)	✅	68.642	76.409	81.242	17
[boosttrack]	                                ✅	67.683	75.764	80.181	
[bytetrack](https://arxiv.org/abs/2110.06864)	✅	66.773	76.929	77.874	1265
[ocsort](https://arxiv.org/abs/2203.14360)	✅	65.469	74.934	75.983	1483
[imprassoc]	                                ✅	64.36	76.578	71.903	26
[deepocsort](https://arxiv.org/abs/2302.11813)	✅	63.18	74.474	73.496	12
[hybridsort](https://arxiv.org/abs/2308.00783)	❌
```
Without ECC
```
                                      Tracker	Status	HOTA↑	MOTA↑	IDF1↑	FPS
[botsort](https://arxiv.org/abs/2206.14651)	✅	68.774	77.176	80.997	46
[strongsort](https://arxiv.org/abs/2202.13514)	✅	68.642	76.409	81.242	17
[boosttrack]	                                ✅	66.894	75.413	78.351
[bytetrack](https://arxiv.org/abs/2110.06864)	✅	66.773	76.929	77.874	1265
[ocsort](https://arxiv.org/abs/2203.14360)	✅	65.469	74.934	75.983	1483
[imprassoc]	                                ✅	64.36	76.578	71.903	26
[deepocsort](https://arxiv.org/abs/2302.11813)	✅	63.18	74.474	73.496	12
[hybridsort](https://arxiv.org/abs/2308.00783)	❌
```
A bit better than ByteTrack still