@Fleyderer 2025-03-06
With my detections, which gives closest to original boosttrack results, without ecc bytetrack is better. I am a bit tired of these small differences in metrics, which are more a matter of chance than of real advantage :smile:

But what about FPS without ECC? I see that FPS here is the same. 

BTW, what do you think about adding + and ++ versions? I am currently working on it, and I would like to know how it will be included in the boxmot. I think, it would be better to add this tracker as 'boosttrack' and let people decide which parameters they need to use or not. And use the BoostTrack++ config by default (when it will be ready).