@mikel-brostrom 2025-03-06
> Of course, main contribution of this tracker is BoostTrack++ version, but currently I don't know how to better add both + and ++ to boxmot. There are only few changes in configs, and now there is no functionality in boxmot to add tracker without changing its class while changing only its config file. This thing should be discussed and maybe you have your own opinion.

My usual way of working regarding this is that the config under the repo represents the maximum tracker performance (++). Maybe a comment in the config is enough with the `regular`, `+` and `++` configurations?