@mikel-brostrom 2025-03-06
`I think, it would be better to add this tracker as 'boosttrack' and let people decide which parameters they need to use or not. And use the BoostTrack++ config by default (when it will be ready).`

Exactly 🚀