@Fleyderer 2025-03-06
> I think, it would be better to add this tracker as 'boosttrack' and let people decide which parameters they need to use or not. And use the BoostTrack++ config by default (when it will be ready).

> My usual way of working regarding this is that the config under the repo represents the maximum tracker performance (++). Maybe a comment in the config is enough with the `regular`, `+` and `++` configurations

great minds think alike :smile: