@Fleyderer 2025-03-06
Succesfully added BoostTrack+ and ++ versions.

Comments:

1. Fixed normalization in `base_backend.py` to normalize each vector independently. In most cases, when cos_sim is used, it does not matter, but in boosttrack, where `emb_cost = dets_embs @ tracks_embs` it would be incorrect.
2. Added fix #1793 back, idk why did it gone from master :smile: 

### 3. Reproduce 

Original BoostTrack: `{"HOTA": 68.486, "MOTA": 75.508, "IDF1": 81.277}`

BoostTrack+: `{"HOTA": 69.237, "MOTA": 75.623, "IDF1": 82.696}`
BoostTrack++: `{"HOTA": 69.254, "MOTA": 76.196, "IDF1": 82.844}`

Boxmot BoostTrack  with `osnet_x1_0_msmt17.pt`:
BoostTrack+: `{"HOTA": 69.074, "MOTA": 75.554, "IDF1": 82.033}`
BoostTrack++: `{"HOTA": 69.177, "MOTA": 76.146, "IDF1": 82.485}`

Boxmot BoostTrack with original BoostTrack ReID:
BoostTrack+: `{"HOTA": 69.007, "MOTA": 75.546, "IDF1": 81.936}`
BoostTrack++: `{"HOTA": 69.229, "MOTA": 76.255, "IDF1": 82.333}`

How to add original ReID:

1. Change NR_CLASSES_DICT `reid_model_factory.py`:

```
NR_CLASSES_DICT = {
    'market1501': 751,
    'duke': 702,
    'veri': 576,
    'vehicleid': 576,
    'boost': 2510
}
```

2. Download weights from [here](https://drive.google.com/file/d/1RCV1ZxzZCjGKV-kYyZaTIwG4luoIfa7B/view?usp=drive_link), put it in boxmot directory and rename from `osnet_ain_ms_d_c.pth.tar` to `osnet_ain_x1_0_boost.pt`.