@Fleyderer 2025-03-06
One more thing: does the new FPS feature take into account the calculation of embeddings in online mode, and not pre-calculated ones?