@mikel-brostrom 2025-03-06
Not pre-calculated. At the moment it is just a print in one of the tests, so it has to be added to the README by hand. It takes the average of 100 updates, not counting the first warmup iteration.