@mikel-brostrom 2025-03-06
Hmmm. The benchmark gave very low results