@Fleyderer 2025-03-07
> Hmmm. The benchmark gave very low results for boosttrack : https://github.com/mikel-brostrom/boxmot/pull/1849/files. Could you have a look at it?

I am trying to reproduce results step by step:

Firstly, I have found [mot-metrics-benchmark](https://github.com/mikel-brostrom/boxmot/blob/210e4d0f2c1f80c39aa571c1f89b65e897890c22/.github/workflows/benchmark.yml#L21)

Secondly, I've downloaded data using steps from this job:

```
wget https://github.com/mikel-brostrom/boxmot/releases/download/v10.0.83/MOT17-50.zip -O tracking/val_utils/MOT17-50.zip
unzip tracking/val_utils/MOT17-50.zip -d ./tracking/val_utils/data/
```

Thirdly, ran val.py:

```
python3 tracking/val.py --imgsz 320 --classes 0 --yolo-model yolov8x.pt --reid-model osnet_x1_0_dukemtmcreid.pt --tracking-method boosttrack --verbose --source ./tracking/val_utils/data/MOT17-50/train
```

And got some errors:

```
2025-03-07 09:14:38.698 | ERROR    | tracking.utils:download_mot_eval_tools:135 - Error processing /home/kozakmv/Programming/Python/boxmot/tracking/val_utils/data/__MACOSX/MOT17-50/train/MOT17-05/gt/._gt.txt: 'utf-8' codec can't decode byte 0xa9 in position 37: invalid start byte
```

And also results are:

```
{"HOTA": 33.781, "MOTA": 22.394, "IDF1": 35.11}
```

For bytetrack results are:

```
{"HOTA": 33.661, "MOTA": 23.004, "IDF1": 35.442}
```

These results are not even close to HOTA 68+.  I want to reproduce results by myself, without already existing dets_n_embs cache, because there were problems with embeddings normalization etc.