@mikel-brostrom 2025-03-07
Benchmark here: https://github.com/mikel-brostrom/boxmot/blob/master/.github/workflows/benchmark.yml 😄